Return to scaling before saving

+++ scripts/train.py
@@ -67,7 +67,7 @@ def main():
 
     os.makedirs(output_folder, exist_ok=True)
     dataset_path = f'{output_folder}/train/datasets/{model_name}'
-    # scaled_dataset_path = f'{output_folder}/train/datasets/{model_name}_scaled'
+    scaled_dataset_path = f'{output_folder}/train/datasets/{model_name}_scaled'
 
     model_folder = f'{output_folder}/models/{model_name}'
     stats_folder = f'{output_folder}/stats/{model_name}'
@@ -92,8 +92,8 @@ def main():
         os.makedirs(output_temp, exist_ok=True)
         
         # same for scaled datasets
-        # output_temp_scaled = f'{output_folder}/train/temp_scaled/'
-        # os.makedirs(output_temp_scaled, exist_ok=True)
+        output_temp_scaled = f'{output_folder}/train/temp_scaled/'
+        os.makedirs(output_temp_scaled, exist_ok=True)
 
         # multiprocess build of connections
         pool_obj = ThreadPool()
@@ -161,42 +161,49 @@ def main():
 
         # applied transformations over all intermediate path light dataset
         # avoid memory overhead
-        # if not os.path.exists(f'{scaled_dataset_path}.train'):
+        if not os.path.exists(f'{scaled_dataset_path}.train'):
             
-        #     intermediate_scaled_datasets = []
-        #     intermediate_datasets_path = sorted(os.listdir(output_temp))
+            intermediate_scaled_datasets = []
+            intermediate_datasets_path = sorted(os.listdir(output_temp))
             
-        #     n_subsets = len(intermediate_datasets_path)
-        #     step = (n_subsets // 100) + 1
+            n_subsets = len(intermediate_datasets_path)
+            step = (n_subsets // 100) + 1
             
-        #     for idx, dataset_name in enumerate(intermediate_datasets_path):
-        #         c_dataset_path = os.path.join(output_temp, dataset_name)
-        #         c_dataset = PathLightDataset(root=c_dataset_path)
+            for idx, dataset_name in enumerate(intermediate_datasets_path):
+                c_dataset_path = os.path.join(output_temp, dataset_name)
+                c_dataset = PathLightDataset(root=c_dataset_path)
                 
-        #         c_scaled_dataset_path = os.path.join(output_temp_scaled, dataset_name)
-        #         scaled_dataset = PathLightDataset(c_scaled_dataset_path, c_dataset, pre_transform=applied_transforms)
+                c_scaled_dataset_path = os.path.join(output_temp_scaled, dataset_name)
+                scaled_dataset = PathLightDataset(c_scaled_dataset_path, c_dataset, 
+                                                pre_transform=applied_transforms)
                 
-        #         if (idx % step == 0 or idx >= n_subsets - 1):
-        #             print(f'[Scaling] -- progress: {(idx + 1) / n_subsets * 100.:.2f}%', \
-        #                 end='\r' if idx + 1 < n_subsets else '\n')
+                if (idx % step == 0 or idx >= n_subsets - 1):
+                    print(f'[Scaling] -- progress: {(idx + 1) / n_subsets * 100.:.2f}%', \
+                        end='\r' if idx + 1 < n_subsets else '\n')
                 
-        #         # now transform dataset using scaler and encoding
-        #         intermediate_scaled_datasets.append(scaled_dataset)
+                # now transform dataset using scaler and encoding
+                intermediate_scaled_datasets.append(scaled_dataset)
+                del c_dataset # remove intermediate variable
 
-        #     scaled_concat_datasets = torch.utils.data.ConcatDataset(intermediate_scaled_datasets)
-        #     scaled_concatenated_dataset = PathLightDataset(scaled_dataset_path, scaled_concat_datasets)
+            scaled_concat_datasets = torch.utils.data.ConcatDataset(intermediate_scaled_datasets)
+            scaled_concatenated_dataset = PathLightDataset(scaled_dataset_path, 
+                                                    scaled_concat_datasets)
         
             # save scaled dataset
-        print(f'Save train and test dataset into: {dataset_path}')
-        PathLightDataset(f'{dataset_path}.train', 
-                        train_dataset, transform=applied_transforms)
-        PathLightDataset(f'{dataset_path}.test', 
-                        test_dataset, transform=applied_transforms)
-    
-        print('[cleaning] clear intermediates saved datasets...')
-        os.system(f'rm -r {output_temp}')
-            # os.system(f'rm -r {output_temp_scaled}')
-            # os.system(f'rm -r {scaled_dataset_path}') # remove also previous computed dataset
+            print(f'Save train and test dataset into: {dataset_path}')
+            PathLightDataset(f'{dataset_path}.train', 
+                            scaled_concatenated_dataset[:split_index])
+            PathLightDataset(f'{dataset_path}.test', 
+                            scaled_concatenated_dataset[split_index:])
+        
+            print('[cleaning] clear intermediates saved datasets...')
+            os.system(f'rm -r {output_temp}')
+            os.system(f'rm -r {output_temp_scaled}')
+            os.system(f'rm -r {scaled_dataset_path}') # remove also previous computed dataset
+            
+            # remove intermediate variables
+            del scaled_concat_datasets
+            del scaled_concatenated_dataset
 
     # need to reload scalers and transformers?
     x_scaler = skload(f'{model_folder}/x_node_scaler.bin')
@@ -220,10 +227,12 @@ def main():
     
     # TODO: use of GPU based dataset?
     if train_dataset is None or test_dataset is None:
-        train_dataset = PathLightDataset(root=f'{dataset_path}.train', transform=applied_transforms)
-        test_dataset = PathLightDataset(root=f'{dataset_path}.test', transform=applied_transforms)
-    print(f'Example of scaled element from train dataset: {train_dataset[0]}')
-
+        train_dataset = PathLightDataset(root=f'{dataset_path}.train')
+        test_dataset = PathLightDataset(root=f'{dataset_path}.test')
+        
+    print(f'Train dataset: {len(train_dataset)} graphs')
+    print(f'Test dataset: {len(test_dataset)} graphs')
+    print(f'Example of graph from train dataset: {train_dataset[0]}')
     train_loader = DataLoader(train_dataset, batch_size=MIGNNConf.BATCH_SIZE, shuffle=True)
     test_loader = DataLoader(test_dataset, batch_size=MIGNNConf.BATCH_SIZE, shuffle=True)
